// src/serverApi/v3/models/school-external-tool-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * HPI Schul-Cloud Server API
 * This is v3 of HPI Schul-Cloud Server. Checkout /docs for v1.
 *
 * The version of the OpenAPI document: 3.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { CustomParameterEntryResponse } from './custom-parameter-entry-response';

/**
 * 
 * @export
 * @interface SchoolExternalToolResponse
 */
export interface SchoolExternalToolResponse {
    /**
     * 
     * @type {string}
     * @memberof SchoolExternalToolResponse
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof SchoolExternalToolResponse
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof SchoolExternalToolResponse
     */
    toolId: string;
    /**
     * 
     * @type {string}
     * @memberof SchoolExternalToolResponse
     */
    schoolId: string;
    /**
     * 
     * @type {Array<CustomParameterEntryResponse>}
     * @memberof SchoolExternalToolResponse
     */
    parameters: Array<CustomParameterEntryResponse>;
    /**
     * 
     * @type {number}
     * @memberof SchoolExternalToolResponse
     */
    toolVersion: number;
    /**
     * 
     * @type {string}
     * @memberof SchoolExternalToolResponse
     */
    status: SchoolExternalToolResponseStatus;
}

/**
    * @export
    * @enum {string}
    */
export enum SchoolExternalToolResponseStatus {
    LATEST = 'Latest',
    OUTDATED = 'Outdated',
    UNKNOWN = 'Unknown'
}